+++ src/main/kotlin/io/hhplus/tdd/point/controller/PointController.kt
@@ -1,5 +1,7 @@
-package io.hhplus.tdd.point
+package io.hhplus.tdd.point.controller
 
+import io.hhplus.tdd.point.entity.PointHistory
+import io.hhplus.tdd.point.entity.UserPoint
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.web.bind.annotation.*
