+++ src/main/kotlin/io/hhplus/tdd/point/entity/PointHistory.kt
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point
+package io.hhplus.tdd.point.entity
 
 data class PointHistory(
     val id: Long,
